Deprecate every kind of configuration
Rule authors do not need to know about internal implementations of YamlConfig etc.
They know that the yaml format config is supported and what methods can be used with a Config object.
