+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-base-config/-init-.md
@@ -1,0 +1,12 @@
+---
+title: BaseConfig.<init> - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [BaseConfig](index.html) / [&lt;init&gt;](./-init-.html)
+
+# &lt;init&gt;
+
+`BaseConfig()`
+
+Convenient base configuration which parses/casts the configuration value based on the type of the default value.
+
