+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-base-config/try-parse-based-on-default.md
@@ -1,0 +1,9 @@
+---
+title: BaseConfig.tryParseBasedOnDefault - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [BaseConfig](index.html) / [tryParseBasedOnDefault](./try-parse-based-on-default.html)
+
+# tryParseBasedOnDefault
+
+`protected open fun tryParseBasedOnDefault(result: `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`, defaultResult: `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`): `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)
