+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-base-config/value-or-default-internal.md
@@ -1,0 +1,9 @@
+---
+title: BaseConfig.valueOrDefaultInternal - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [BaseConfig](index.html) / [valueOrDefaultInternal](./value-or-default-internal.html)
+
+# valueOrDefaultInternal
+
+`protected open fun valueOrDefaultInternal(key: `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`, result: `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`?, default: `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`): `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)
