+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-composite-config/-init-.md
@@ -1,0 +1,12 @@
+---
+title: CompositeConfig.<init> - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [CompositeConfig](index.html) / [&lt;init&gt;](./-init-.html)
+
+# &lt;init&gt;
+
+`CompositeConfig(lookFirst: `[`Config`](../../io.gitlab.arturbosch.detekt.api/-config/index.html)`, lookSecond: `[`Config`](../../io.gitlab.arturbosch.detekt.api/-config/index.html)`)`
+
+Wraps two different configuration which should be considered when retrieving properties.
+
