+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-composite-config/to-string.md
@@ -1,0 +1,9 @@
+---
+title: CompositeConfig.toString - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [CompositeConfig](index.html) / [toString](./to-string.html)
+
+# toString
+
+`fun toString(): `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)
