+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-composite-config/value-or-null.md
@@ -1,0 +1,13 @@
+---
+title: CompositeConfig.valueOrNull - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [CompositeConfig](index.html) / [valueOrNull](./value-or-null.html)
+
+# valueOrNull
+
+`fun <T : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> valueOrNull(key: `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`): T?`
+
+Retrieves a sub configuration or value based on given key.
+If the configuration property cannot be found, null is returned.
+
