+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-yaml-config/load-resource.md
@@ -1,0 +1,12 @@
+---
+title: YamlConfig.loadResource - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [YamlConfig](index.html) / [loadResource](./load-resource.html)
+
+# loadResource
+
+`fun loadResource(url: `[`URL`](https://docs.oracle.com/javase/8/docs/api/java/net/URL.html)`): `[`Config`](../../io.gitlab.arturbosch.detekt.api/-config/index.html)
+
+Factory method to load a yaml configuration from a URL.
+
