+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-yaml-config/parent.md
@@ -1,0 +1,12 @@
+---
+title: YamlConfig.parent - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [YamlConfig](index.html) / [parent](./parent.html)
+
+# parent
+
+`val parent: Parent?`
+
+Returns the parent config which encloses this config part.
+
