+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/-yaml-config/sub-config.md
@@ -1,0 +1,12 @@
+---
+title: YamlConfig.subConfig - detekt-api
+---
+
+[detekt-api](../../index.html) / [io.gitlab.arturbosch.detekt.api.internal](../index.html) / [YamlConfig](index.html) / [subConfig](./sub-config.html)
+
+# subConfig
+
+`fun subConfig(key: `[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`): `[`Config`](../../io.gitlab.arturbosch.detekt.api/-config/index.html)
+
+Tries to retrieve part of the configuration based on given key.
+
