+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api/-base-config.md
@@ -1,0 +1,13 @@
+---
+title: BaseConfig - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api](index.html) / [BaseConfig](./-base-config.html)
+
+# BaseConfig
+
+`typealias ~~BaseConfig~~ = `[`BaseConfig`](../io.gitlab.arturbosch.detekt.api.internal/-base-config/index.html)
+**Deprecated:** 'BaseConfig' exposes implementation details of 'YamlConfig' and should't be relied on.
+
+Convenient base configuration which parses/casts the configuration value based on the type of the default value.
+
