+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api/-composite-config.md
@@ -1,0 +1,13 @@
+---
+title: CompositeConfig - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api](index.html) / [CompositeConfig](./-composite-config.html)
+
+# CompositeConfig
+
+`typealias ~~CompositeConfig~~ = `[`CompositeConfig`](../io.gitlab.arturbosch.detekt.api.internal/-composite-config/index.html)
+**Deprecated:** Rule authors should not rely on a specific configuration kind.
+
+Wraps two different configuration which should be considered when retrieving properties.
+
