+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api/-yaml-config.md
@@ -1,0 +1,14 @@
+---
+title: YamlConfig - detekt-api
+---
+
+[detekt-api](../index.html) / [io.gitlab.arturbosch.detekt.api](index.html) / [YamlConfig](./-yaml-config.html)
+
+# YamlConfig
+
+`typealias ~~YamlConfig~~ = `[`YamlConfig`](../io.gitlab.arturbosch.detekt.api.internal/-yaml-config/index.html)
+**Deprecated:** Rule authors should not rely on a specific configuration kind.
+
+Config implementation using the yaml format. SubConfigurations can return sub maps according to the
+yaml specification.
+
